#159: 파노라마 뷰컨트롤러와 익스텐션을 하나로 병합
